requirements.txt
pytest
requests
python-dotenv
pytest-playwright